+++ demo/index.html
@@ -19,7 +19,6 @@
     <script src="//cdnjs.cloudflare.com/ajax/libs/es5-shim/4.1.15/es5-shim.min.js"></script>
     <script src="//cdnjs.cloudflare.com/ajax/libs/es5-shim/4.1.15/es5-sham.min.js"></script>
   <![endif]-->
-  <script src="http://localhost:3000/webpack-dev-server.js"></script>
   <script src="assets/main.js"></script>
   <script>
     // Sanity-check the component loaded...
